rt: auto-impl parameters
