+++ librashader-runtime-metal/src/lib.rs
@@ -1,3 +1,4 @@
+#![cfg(target_vendor = "apple")]
 mod draw_quad;
 mod error;
 mod graphics_pipeline;
